[refractor] replace ConstraintLayout with ScrollView and rename files#1
