+++ app/src/main/java/com/example/rewear/ui/my/MyFragment.kt
@@ -1,4 +1,4 @@
-package com.example.rewear.ui.notifications
+package com.example.rewear.ui.my
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,11 +7,11 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.example.rewear.databinding.FragmentNotificationsBinding
+import com.example.rewear.databinding.FragmentMyBinding
 
-class NotificationsFragment : Fragment() {
+class MyFragment : Fragment() {
 
-    private var _binding: FragmentNotificationsBinding? = null
+    private var _binding: FragmentMyBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -22,14 +22,14 @@ class NotificationsFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        val notificationsViewModel =
-            ViewModelProvider(this).get(NotificationsViewModel::class.java)
+        val myViewModel =
+            ViewModelProvider(this).get(MyViewModel::class.java)
 
-        _binding = FragmentNotificationsBinding.inflate(inflater, container, false)
+        _binding = FragmentMyBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textNotifications
-        notificationsViewModel.text.observe(viewLifecycleOwner) {
+        val textView: TextView = binding.textMy
+        myViewModel.text.observe(viewLifecycleOwner) {
             textView.text = it
         }
         return root
