+++ app/src/main/java/com/example/rewear/ui/my/MyViewModel.kt
@@ -1,10 +1,10 @@
-package com.example.rewear.ui.notifications
+package com.example.rewear.ui.my
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class NotificationsViewModel : ViewModel() {
+class MyViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is notifications Fragment"
