+++ app/src/main/java/com/example/rewear/ui/reform/ReformFragment.kt
@@ -1,4 +1,4 @@
-package com.example.rewear.ui.dashboard
+package com.example.rewear.ui.reform
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,11 +7,11 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.example.rewear.databinding.FragmentDashboardBinding
+import com.example.rewear.databinding.FragmentReformBinding
 
-class DashboardFragment : Fragment() {
+class ReformFragment : Fragment() {
 
-    private var _binding: FragmentDashboardBinding? = null
+    private var _binding: FragmentReformBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -22,14 +22,14 @@ class DashboardFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        val dashboardViewModel =
-            ViewModelProvider(this).get(DashboardViewModel::class.java)
+        val reformViewModel =
+            ViewModelProvider(this).get(ReformViewModel::class.java)
 
-        _binding = FragmentDashboardBinding.inflate(inflater, container, false)
+        _binding = FragmentReformBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textDashboard
-        dashboardViewModel.text.observe(viewLifecycleOwner) {
+        val textView: TextView = binding.textReform
+        reformViewModel.text.observe(viewLifecycleOwner) {
             textView.text = it
         }
         return root
