+++ app/src/main/java/com/example/rewear/ui/reform/ReformViewModel.kt
@@ -1,10 +1,10 @@
-package com.example.rewear.ui.dashboard
+package com.example.rewear.ui.reform
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class DashboardViewModel : ViewModel() {
+class ReformViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is dashboard Fragment"
